엔티티 추가 및 수정, CountRepository 추가, 모든 id -> Integer로 변경

+++ src/main/java/com/example/demo/Controller/AuthController.java
@@ -21,7 +21,7 @@ public class AuthController {
     public String login(@RequestBody LoginRequestDto loginDto,
                         HttpServletRequest request){
 
-        Long userId = userService.login(loginDto.getEmail(), loginDto.getPassword());
+        Integer userId = userService.login(loginDto.getEmail(), loginDto.getPassword());
 
 
         HttpSession session = request.getSession(true);
