+++ src/main/java/com/example/demo/Repository/PostRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post,Integer> {
 
-    Slice<Post> findByIdLessThanOrderByIdDesc(Long lastPostId, Pageable pageable);
+    Slice<Post> findByIdLessThanOrderByIdDesc(Integer lastPostId, Pageable pageable);
 }
