+++ src/main/java/com/example/demo/Repository/ReplyRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.Repository;
 import com.example.demo.Entity.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReplyRepository extends JpaRepository<Reply,Long> {
+public interface ReplyRepository extends JpaRepository<Reply,Integer> {
 }
